requirements.txt
python-telegram-bot==20.7
psycopg[binary]==3.1.18